Add prettier output messages for validation serve responses
