+++ tests/unit/main.test.ts
@@ -1,6 +1,6 @@
-import { getInput, setFailed } from "@actions/core";
+import { error, getInput, setFailed } from "@actions/core";
 import axios from "axios";
-import { main } from "../../src";
+import { main } from "../../src/main";
 
 jest.mock('@actions/github', () => {
 	const originalModule = jest.requireActual('@actions/github');
@@ -17,22 +17,46 @@ jest.mock('@actions/github', () => {
 	};
 });
 
+const badValidationResponse = {
+	status: 200,
+	statusText: "OK",
+	headers: {},
+	config: {},
+	data: {
+		validationMessages: [
+			{
+				fieldPath: "path",
+				description: "failed!",
+				validationCode: "01",
+			}
+		]
+	}
+};
+
+const goodValidationResponse = {
+	status: 200,
+	statusText: "OK",
+	headers: {},
+	config: {},
+	data: {
+		validationMessages: []
+	}
+};
+
 jest.mock('@actions/core');
 jest.mock("axios");
 
 describe("main", () => {
 	it("should call setFailed if validation fails", async () => {
-		axios.post.mockReturnValue({
-			status: 200,
-			statusText: "OK",
-			headers: {},
-			config: {},
-			data: {
-				validationMessages: ["failed!"]
-			}
-		})
+		axios.post.mockReturnValue(badValidationResponse);
 		await main();
 		expect(setFailed).toBeCalled();
+	})
+	it("should output errors if validation fails", async () => {
+		axios.post.mockReturnValue(badValidationResponse);
+		await main();
+		expect(error).toBeCalled();
+		expect(error).toBeCalledWith(expect.stringContaining("failed!"));
 	})
 	it("should call setFailed if an error occurs anywhere", async () => {
 		axios.post.mockImplementation(() => {throw new Error("Mocked error")});
@@ -40,31 +64,22 @@ describe("main", () => {
 		expect(setFailed).toBeCalled();
 	})
 	it("should not call setFailed if everything works as expected", async () => {
-		axios.post.mockReturnValue({
-			status: 200,
-			statusText: "OK",
-			headers: {},
-			config: {},
-			data: {
-				validationMessages: []
-			}
-		})
+		axios.post.mockReturnValue(goodValidationResponse);
 		await main();
 		expect(setFailed).toBeCalledTimes(0);
 	})
-	it("should only send requests to validation server if dry-run is enabled", async () => {
-		axios.post.mockReturnValue({
-			status: 200,
-			statusText: "OK",
-			headers: {},
-			config: {},
-			data: {
-				validationMessages: []
-			}
-		})
+	it("should send requests to validation server only, when dry-run is enabled", async () => {
+		axios.post.mockReturnValue(goodValidationResponse);
 		getInput.mockReturnValue("true");
 		await main();
 		expect(axios.post).toBeCalledTimes(1);
 		expect(axios.post).toBeCalledWith(expect.stringContaining("debug"), expect.anything(), expect.anything());
+	});
+
+	it("should output errors when validation fails, when dry-run is enabled", async () => {
+		axios.post.mockReturnValue(badValidationResponse);
+		await main();
+		expect(error).toBeCalled();
+		expect(error).toBeCalledWith(expect.stringContaining("failed!"));
 	});
 })
